Added tests for repo + parser
