+++ src/test/kotlin/com/discounts/nearby/repository/SupermarketRepositoryTest.kt
@@ -30,7 +30,29 @@ class SupermarketRepositoryTest {
 
     @Test
     fun `find entity by id`() {
-        val toPersist = Supermarket().apply {
+        val persisted = prepareEntity()
+        val found = supermarketRepository.findByIdOrNull(persisted.id)
+
+        assertNotNull(found)
+        assertEquals(persisted, found)
+    }
+
+    @Test
+    fun `find entity by its code`() {
+        val persisted = prepareEntity()
+        val found = supermarketRepository.getSupermarketByCode(SupermarketCode.OKEY)
+
+        assertNotNull(found)
+        assertEquals(persisted, found)
+    }
+
+    private fun prepareEntity(entity: Supermarket = ENTITY_TO_PERSIST) =
+        entityManager
+            .merge(entity)
+            .also { entityManager.flush() }
+
+    private companion object {
+        val ENTITY_TO_PERSIST = Supermarket().apply {
             name = "name"
             code = SupermarketCode.OKEY
             goodsSortedByPrice = Goods(
@@ -48,14 +70,5 @@ class SupermarketRepositoryTest {
                 goods = emptyList()
             )
         }
-
-        val persisted = entityManager
-            .persist(toPersist)
-            .also { entityManager.flush() }
-
-        val found = supermarketRepository.findByIdOrNull(persisted.id)
-
-        assertNotNull(found)
-        assertEquals(persisted, found)
     }
 }
