+++ src/test/kotlin/com/discounts/nearby/service/parse/LentaSiteDataParserTest.kt
@@ -1,10 +1,10 @@
-package com.discounts.nearby.service
+package com.discounts.nearby.service.parse
 
 import com.discounts.nearby.model.SupermarketCode
 import com.discounts.nearby.model.category.GoodCategory
-import com.discounts.nearby.model.category.OkeyCategory
+import com.discounts.nearby.model.category.LentaCategory
 import com.discounts.nearby.service.supermarket.category.manager.SupermarketCategoryManager
-import com.discounts.nearby.service.supermarket.parser.provider.impl.OkeySiteDataProvider
+import com.discounts.nearby.service.supermarket.parser.provider.impl.LentaSiteDataProvider
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
@@ -17,17 +17,17 @@ import org.mockito.junit.jupiter.MockitoExtension
  * @author shvatov
  */
 @ExtendWith(MockitoExtension::class)
-internal class OkeySiteDataParserTest {
+internal class LentaSiteDataParserTest {
     @Mock
     private lateinit var goodCategoryManager: SupermarketCategoryManager
 
     @InjectMocks
-    private lateinit var service: OkeySiteDataProvider
+    private lateinit var service: LentaSiteDataProvider
 
     @Test
     fun `test parsing the website - all elements`() {
-        `when`(goodCategoryManager.getLocalized(SupermarketCode.OKEY, GoodCategory.BEER))
-            .thenReturn(OkeyCategory.BEER.localizedCategory)
+        `when`(goodCategoryManager.getLocalized(SupermarketCode.LENTA, GoodCategory.BEER))
+            .thenReturn(LentaCategory.BEER.localizedCategory)
 
         val goods = service.getDataByCategory(GoodCategory.BEER, 10, false)
         assertTrue { goods.goods?.size != 0 }
@@ -35,16 +35,10 @@ internal class OkeySiteDataParserTest {
 
     @Test
     fun `test parsing the website - discounts only`() {
-        `when`(goodCategoryManager.getLocalized(SupermarketCode.OKEY, GoodCategory.BEER))
-            .thenReturn(OkeyCategory.BEER.localizedCategory)
+        `when`(goodCategoryManager.getLocalized(SupermarketCode.LENTA, GoodCategory.BEER))
+            .thenReturn(LentaCategory.BEER.localizedCategory)
 
         val goods = service.getDataByCategory(GoodCategory.BEER, 10, true)
-        assertTrue { goods.goods?.size != 0 }
-    }
-
-    @Test
-    fun `test parsing the website - find by key words`() {
-        val goods = service.getDataByGoodName("сок добрый", 72)
         assertTrue { goods.goods?.size != 0 }
     }
 }
